wip refactoring to extract database code
