+++ EmpowerPlant/EmpowerPlantViewController.swift
@@ -9,10 +9,6 @@ import UIKit
 import Sentry
 
 class EmpowerPlantViewController: UIViewController {
-    
-    // CoreData database
-    let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-    
     let tableView: UITableView = {
         let table = UITableView(frame: .zero, style: .plain)
         table.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
@@ -36,24 +32,14 @@ class EmpowerPlantViewController: UIViewController {
             tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             tableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
         ])
-        
-        // Configures the nav bar buttons
+
         configureNavigationItems()
-        
-        // ???: looks like this was already done?
-        /* TODO: implement:
-         1 get products from server (so we get http.client span)
-         2 check if any products in Core Data -> If Not -> insert the products from response into Core Data
-         3 get products from DB (so we get db.query span) and reload the table with this data
-         */
-        
         getAllProductsFromServer()
         getAllProductsFromDb()
         readCurrentDirectory()
         performLongFileOperation()
         processProducts()
         checkRelease()
-        
         
         NotificationCenter.default.addObserver(forName: modifiedDBNotificationName, object: nil, queue: nil) { _ in
             self.getAllProductsFromDb()
@@ -125,13 +111,10 @@ class EmpowerPlantViewController: UIViewController {
     }
     
     func fibonacciSeries(num: Int) -> Int{
-        // The value of 0th and 1st number of the fibonacci series are 0 and 1
         var n1 = 0
         var n2 = 1
-        
-        // To store the result
+
         var nR = 0
-        // Adding two previous numbers to find ith number of the series
         for _ in 0..<num{
             nR = n1
             n1 = n2
@@ -155,29 +138,20 @@ class EmpowerPlantViewController: UIViewController {
             guard let field = alert.textFields?.first, let text = field.text, !text.isEmpty else {
                 return
             }
-            self?.createProduct(productId: "123", title: text, productDescription: "product.description", productDescriptionFull: "product.description.full", img:"img", imgCropped:"img.cropped", price:"1")
+            CoreDataController.shared.createProduct(productId: "123", title: text, productDescription: "product.description", productDescriptionFull: "product.description.full", img:"img", imgCropped:"img.cropped", price:"1")
         }))
         
         self.present(alert, animated: true, completion: nil)
-        
-        // ALSO WORKED
-        // alert.addTextField()
-        // let submitButton = UIAlertAction(title:"Add", style: .default) { (action) in
-        //     print("here")
-        //     let textfield = alert.textFields![0]
-        // }
-        // alert.addAction(submitButton)
-        // self.present(alert, animated: true, completion: nil)
     }
     
-    // Don't deprecate, this function is useful for development and testing
+    // TODO: merge this and and wipeDB
     @objc
     func clearDb() {
         print("> clearDb")
         // self.products was already set by viewDidLoad()
         // self.products = try context.fetch(Product.fetchRequest())
         for product in self.products {
-            deleteProduct(product: product)
+            CoreDataController.shared.deleteProduct(product: product)
         }
         refreshTable()
     }
@@ -187,10 +161,9 @@ class EmpowerPlantViewController: UIViewController {
             image: UIImage(systemName: "cart"),
             style: .plain,
             target: self,
-            action: #selector(goToCart) // addToDb
+            action: #selector(goToCart)
         )
         self.navigationItem.rightBarButtonItem?.accessibilityIdentifier = "Cart"
-        //self.navigationItem.rightBarButtonItem?.badgeValue = "\(1)"
         
         self.navigationItem.leftBarButtonItems = [UIBarButtonItem(
             image: UIImage(systemName: "ellipsis"),
@@ -228,69 +201,16 @@ class EmpowerPlantViewController: UIViewController {
             if totalItems == 0 {
                 totalItems = defaultTotalItems
             }
-            var itemsPerBatch = 1_000
-            let batches = totalItems / itemsPerBatch
-            
-            let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-            DispatchQueue.global(qos: .utility).async {
-                for i in 0..<batches {
-                    DispatchQueue.main.async {
-                        for j in 0..<itemsPerBatch {
-                            let newProduct = Product(context: context)
-                            let productNum = i * itemsPerBatch + j
-                            
-                            newProduct.productId = "Product \(productNum)" // 'id' was a reserved word in swift
-                            newProduct.title = "Product \(productNum)"
-                            newProduct.productDescription = "Description for product \(i)" // 'description' was a reserved word in swift
-                            newProduct.productDescriptionFull = "Full description for product \(productNum)"
-                            newProduct.img = "img"
-                            newProduct.imgCropped = "img.cropped"
-                            newProduct.price = "\(productNum)"
-                        }
-                        
-                        do {
-                            try context.save()
-                            NotificationCenter.default.post(name: modifiedDBNotificationName, object: nil)
-                        } catch {
-                            // TODO: error
-                        }
-                    }
-                    // add a small delay so it doesn't lock up the UI
-                    usleep(100_000) // 100 milliseconds
-                }
-            }
+            CoreDataController.shared.generateDBItems(total: totalItems)
         })
         alert.addAction(UIAlertAction(title: "Cancel", style: .destructive))
         present(alert, animated: true)
     }
     
-    // Writes to CoreData database
-    func createProduct(productId: String, title: String, productDescription: String, productDescriptionFull: String, img: String, imgCropped: String, price: String) {
-        let newProduct = Product(context: context)
-        
-        newProduct.productId = productId // 'id' was a reserved word in swift
-        newProduct.title = title
-        newProduct.productDescription = productDescription // 'description' was a reserved word in swift
-        newProduct.productDescriptionFull = productDescriptionFull
-        newProduct.img = img
-        newProduct.imgCropped = imgCropped
-        newProduct.price = price
-    }
-    
-    // Don't deprecate this until major release of this demo
-    func deleteProduct(product: Product) {
-        context.delete(product)
-        do {
-            try context.save()
-        }
-        catch {
-            // TODO: error
-        }
-    }
-    
     func getAllProductsFromDb() {
         do {
-            self.products = try context.fetch(Product.fetchRequest())
+            self.products = try CoreDataController.shared.getAllProducts()
+            // TODO: make this a debug-level log
             // for product in self.products {
             //     print(product.productId, product.title, product.productDescriptionFull)
             // }
@@ -301,24 +221,13 @@ class EmpowerPlantViewController: UIViewController {
         }
     }
     
-    // Also writes them into database if database is empty
+    /// - note Also writes them into database if we don't yet have any products
     func getAllProductsFromServer() {
         let startTime = Date()
         let urlStr = "https://application-monitoring-flask-dot-sales-engineering-sf.appspot.com/products-join"
         let url = URL(string: urlStr)!
         var request = URLRequest(url: url)
         request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        
-        struct ProductMap: Decodable {
-            let id: Int
-            let title: String
-            let description: String
-            let descriptionfull: String
-            let img: String
-            let imgcropped: String
-            let price: Int
-            // reviews: [{id: 4, productid: 4, rating: 4, customerid: null, description: null, created: String},...]
-        }
         
         let task = URLSession.shared.dataTask(with: url) { data, response, error in
             let endTime = Date()
@@ -330,16 +239,15 @@ class EmpowerPlantViewController: UIViewController {
                         var operations = [BlockOperation]()
                         let saveOp = BlockOperation() {
                             do {
-                                try self.context.save()
+                                try CoreDataController.shared.context.save()
                                 self.getAllProductsFromDb()
                             } catch {
                                 // TODO: error
                             }
                         }
                         for product in productsResponse {
-                            // Writes to CoreData database
                             let addOp = BlockOperation() {
-                                self.createProduct(productId: String(product.id), title: product.title, productDescription: product.description, productDescriptionFull: product.descriptionfull, img: product.img, imgCropped: product.imgcropped, price: String(product.price))
+                                CoreDataController.shared.createProduct(product: product)
                             }
                             operations.append(addOp)
                             saveOp.addDependency(addOp)
@@ -391,7 +299,6 @@ extension EmpowerPlantViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let model = products[indexPath.row]
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        //  'model' has all available attributes here, if needed in the future (e.g. UI development)
         cell.textLabel?.text = model.title
         return cell
     }
@@ -403,7 +310,6 @@ extension EmpowerPlantViewController: UITableViewDataSource {
 
 // MARK: UITableViewDelegate
 extension EmpowerPlantViewController: UITableViewDelegate {
-    // Code that executes on Click'ing table row, adds the product item to shopping cart
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let product = products[indexPath.row]
