+++ EmpowerPlant/Utils.swift
@@ -10,7 +10,7 @@ import UIKit
 public let modifiedDBNotificationName = Notification.Name("io.sentry.empowerplants.newly-generated-db-items-available")
 
 public func wipeDB() {
-    guard let url = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.persistentStoreCoordinator.persistentStores.first?.url else {
+    guard let url = CoreDataController.shared.persistentContainer.persistentStoreCoordinator.persistentStores.first?.url else {
         // TODO: error
         return
     }
